提取onEntityItemUpdate方法到抽象类AbstractFilmItem
