+++ src/main/java/com/github/tartaricacid/touhoulittlemaid/inventory/tooltip/ItemMaidTooltip.java
@@ -2,20 +2,5 @@ package com.github.tartaricacid.touhoulittlemaid.inventory.tooltip;
 
 import net.minecraft.world.inventory.tooltip.TooltipComponent;
 
-public class ItemMaidTooltip implements TooltipComponent {
-    private final String modelId;
-    private final String customName;
-
-    public ItemMaidTooltip(String modelId, String customName) {
-        this.modelId = modelId;
-        this.customName = customName;
-    }
-
-    public String getModelId() {
-        return modelId;
-    }
-
-    public String getCustomName() {
-        return customName;
-    }
+public record ItemMaidTooltip(String modelId, String customName) implements TooltipComponent {
 }
