+++ src/main/java/com/github/tartaricacid/touhoulittlemaid/item/AbstractFilmItem.java
@@ -1,0 +1,27 @@
+package com.github.tartaricacid.touhoulittlemaid.item;
+
+import net.minecraft.world.entity.item.ItemEntity;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.phys.Vec3;
+
+public abstract class AbstractFilmItem extends Item {
+    public AbstractFilmItem(Properties pProperties) {
+        super(pProperties);
+    }
+
+    @Override
+    public boolean onEntityItemUpdate(ItemStack stack, ItemEntity entity) {
+        if (!entity.isInvulnerable()) {
+            entity.setInvulnerable(true);
+        }
+        Vec3 position = entity.position();
+        int minY = entity.level.getMinBuildHeight();
+        if (position.y < minY) {
+            entity.setNoGravity(true);
+            entity.setDeltaMovement(Vec3.ZERO);
+            entity.setPos(position.x, minY, position.z);
+        }
+        return super.onEntityItemUpdate(stack, entity);
+    }
+}
