+++ src/main/java/com/github/tartaricacid/touhoulittlemaid/item/ItemPhoto.java
@@ -12,10 +12,8 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.tooltip.TooltipComponent;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.context.UseOnContext;
@@ -27,7 +25,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
-public class ItemPhoto extends Item {
+public class ItemPhoto extends AbstractFilmItem {
     private static final String MAID_INFO = "MaidInfo";
     private static final String CUSTOM_NAME = "CustomName";
 
@@ -77,8 +75,7 @@ public class ItemPhoto extends Item {
 
         // 最后才应用生成实体的逻辑
         Optional<Entity> entityOptional = EntityType.create(getMaidData(photo), worldIn);
-        if (entityOptional.isPresent() && entityOptional.get() instanceof EntityMaid) {
-            EntityMaid maid = (EntityMaid) entityOptional.get();
+        if (entityOptional.isPresent() && entityOptional.get() instanceof EntityMaid maid) {
             maid.setPos(clickLocation.x, clickLocation.y, clickLocation.z);
             // 实体生成必须在服务端应用
             if (!worldIn.isClientSide) {
@@ -90,21 +87,6 @@ public class ItemPhoto extends Item {
         }
 
         return InteractionResult.FAIL;
-    }
-
-    @Override
-    public boolean onEntityItemUpdate(ItemStack stack, ItemEntity entity) {
-        if (!entity.isInvulnerable()) {
-            entity.setInvulnerable(true);
-        }
-        Vec3 position = entity.position();
-        int minY = entity.level.getMinBuildHeight();
-        if (position.y < minY) {
-            entity.setNoGravity(true);
-            entity.setDeltaMovement(Vec3.ZERO);
-            entity.setPos(position.x, minY, position.z);
-        }
-        return super.onEntityItemUpdate(stack, entity);
     }
 
     @Override
