+++ src/main/java/com/github/tartaricacid/touhoulittlemaid/item/ItemSmartSlab.java
@@ -18,15 +18,12 @@ import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.MobSpawnType;
-import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.tooltip.TooltipComponent;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 
@@ -36,7 +33,7 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 
-public class ItemSmartSlab extends Item {
+public class ItemSmartSlab extends AbstractFilmItem {
     private static final String MAID_INFO = "MaidInfo";
     private static final String MAID_OWNER = "Owner";
     private static final String CUSTOM_NAME = "CustomName";
@@ -153,21 +150,6 @@ public class ItemSmartSlab extends Item {
     @Override
     public boolean canFitInsideContainerItems() {
         return this.type != Type.HAS_MAID;
-    }
-
-    @Override
-    public boolean onEntityItemUpdate(ItemStack stack, ItemEntity entity) {
-        if (!entity.isInvulnerable()) {
-            entity.setInvulnerable(true);
-        }
-        Vec3 position = entity.position();
-        int minY = entity.level.getMinBuildHeight();
-        if (position.y < minY) {
-            entity.setNoGravity(true);
-            entity.setDeltaMovement(Vec3.ZERO);
-            entity.setPos(position.x, minY, position.z);
-        }
-        return super.onEntityItemUpdate(stack, entity);
     }
 
     @Override
